+++ mapmemory-rxjava3/src/main/kotlin/internal/SubjectsVisibilityHack.kt
@@ -1,0 +1,7 @@
+@file:Suppress("PackageDirectoryMismatch")
+
+package io.reactivex.rxjava3.subjects
+
+internal fun BehaviorSubject<*>.clear() {
+    setCurrent(null)
+}