+++ CHANGELOG.md
@@ -33,9 +33,8 @@ Many of default accessors are already turned into reusable: `mutableList`, `muta
 
 ### Changes
 
-- Added parameter `initialMap` to `MapMemory.reactiveMutableMap()` extension
-- :warning: **mapmemory-rxjava:** `ReplayStrategy` was removed. Please let us know if this change affected
-  you: https://github.com/RedMadRobot/mapmemory/discussions/20
+- Added parameter `defaultValue` to most of reusable properties: `mutableList`, `mutableMap`, `reactiveMutableMap`, `behaviorSubject`.
+- :warning: **mapmemory-rxjava:** `ReplayStrategy` was removed. Please let us know if this change affected you: https://github.com/RedMadRobot/mapmemory/discussions/20
 
 ### Fixes